api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.20
